+++ test/OpenTelemetry.Exporter.Geneva.Tests/UnixDomainSocketDataTransportTests.cs
@@ -7,7 +7,6 @@ using System.Net.Sockets;
 using System.Reflection;
 using OpenTelemetry.Exporter.Geneva.Transports;
 using OpenTelemetry.Tests;
-using Xunit;
 
 namespace OpenTelemetry.Exporter.Geneva.Tests;