+++ test/OpenTelemetry.Exporter.OneCollector.Tests/OneCollectorIntegrationTests.cs
@@ -9,8 +9,6 @@ using OpenTelemetry.Internal;
 using OpenTelemetry.Logs;
 using OpenTelemetry.Resources;
 using OpenTelemetry.Tests;
-using Xunit;
-using Xunit.Abstractions;
 
 namespace OpenTelemetry.Exporter.OneCollector.Tests;