+++ test/OpenTelemetry.Instrumentation.Http.Tests/HttpClientTests.cs
@@ -10,7 +10,6 @@ using System.Text.Json;
 #endif
 using OpenTelemetry.Metrics;
 using OpenTelemetry.Trace;
-using Xunit;
 
 namespace OpenTelemetry.Instrumentation.Http.Tests;